add section path draw

+++ http/src/main/java/com/angcyo/http/Http.java
@@ -36,6 +36,8 @@ public class Http {
     public static int TIME_OUT = 5_000;
     public static String BASE_URL = "http://www.api.com";
     public static final String TAG = "HttpResult";
+    public static boolean LOG_BODY = BuildConfig.DEBUG;
+    public static boolean LOG_INTERCEPTOR_RESPONSE = BuildConfig.DEBUG;
 
     public static Retrofit.Builder builder(String baseUrl, String logTag) {
         return builder(defaultOkHttpClick(logTag).build(), baseUrl);
@@ -55,6 +57,7 @@ public class Http {
         if (BuildConfig.DEBUG) {
             httpLoggingInterceptorM.setLevel(HttpLoggingInterceptorM.Level.BODY);
         }
+        httpLoggingInterceptorM.logResponse = LOG_INTERCEPTOR_RESPONSE;
         return new OkHttpClient.Builder()
                 .connectTimeout(TIME_OUT, TimeUnit.SECONDS)
                 .readTimeout(TIME_OUT, TimeUnit.SECONDS)
@@ -134,14 +137,14 @@ public class Http {
                                     body = stringResponse.string();
 
                                     //"接口返回数据-->\n" +
-                                    LogUtil.json(TAG, body);
+                                    logJson(TAG, body);
 
                                     if (convertString != null) {
                                         String covert = convertString.covert(body);
                                         if (TextUtils.equals(covert, body)) {
                                             LogUtil.i("IConvertString 转换前后一致");
                                         } else {
-                                            LogUtil.json("转换后", covert);
+                                            logJson("转换后", covert);
                                         }
                                         body = covert;
                                     }
@@ -174,14 +177,14 @@ public class Http {
                                     body = stringResponse.string();
 
                                     //"接口返回数据-->\n" +
-                                    LogUtil.json(TAG, body);
+                                    logJson(TAG, body);
 
                                     if (convert != null) {
                                         String covert = convert.covert(body);
                                         if (TextUtils.equals(covert, body)) {
                                             LogUtil.i("IConvertString 转换前后一致");
                                         } else {
-                                            LogUtil.json("转换后", covert);
+                                            logJson("转换后", covert);
                                         }
                                         body = covert;
                                     }
@@ -222,14 +225,14 @@ public class Http {
                                     body = stringResponse.string();
 
                                     //"接口返回数据-->\n" +
-                                    LogUtil.json(TAG, body);
+                                    logJson(TAG, body);
 
                                     if (convert != null) {
                                         String covert = convert.covert(body);
                                         if (TextUtils.equals(covert, body)) {
                                             LogUtil.i("IConvertString 转换前后一致");
                                         } else {
-                                            LogUtil.json("转换后", covert);
+                                            logJson("转换后", covert);
                                         }
                                         body = covert;
                                     }
@@ -249,6 +252,13 @@ public class Http {
     public static <T> Observable.Transformer<ResponseBody, List<T>> transformerListBean(@NonNull final Class<T> type) {
         return transformerListBean(type, null);
     }
+
+    private static void logJson(String tag, String jsonFormat) {
+        if (LOG_BODY) {
+            LogUtil.json(tag, jsonFormat);
+        }
+    }
+
 
     public interface IConvertString {
         String covert(String body);
@@ -279,6 +289,16 @@ public class Http {
             public void onRemove(String key) {
                 map.remove(key);
             }
+
+            @Override
+            public boolean isKeyAllowEmpty(String key) {
+                return false;
+            }
+
+            @Override
+            public void onEmptyValue(String key, String value) {
+                map.put(key, "");
+            }
         }, args);
         return map;
     }
@@ -302,7 +322,11 @@ public class Http {
                 } else {
                     String value = str.substring(indexOf + 1, length);
                     if (TextUtils.isEmpty(value)) {
-                        onPutValue.onRemove(key);
+                        if (onPutValue.isKeyAllowEmpty(key)) {
+                            onPutValue.onEmptyValue(key, value);
+                        } else {
+                            onPutValue.onRemove(key);
+                        }
                     } else {
                         onPutValue.onValue(key, str.substring(key.length() + 1));
                     }
@@ -326,10 +350,19 @@ public class Http {
                 .build();
     }
 
+    public static RequestBody fileForm(String fileFormKey,
+                                       String filePath,
+                                       String... otherValues) {
+        return fileForm(fileFormKey, filePath, null, otherValues);
+    }
+
     /**
      * 表单形式上传文件, 和其他参数
      */
-    public static RequestBody fileForm(String fileFormKey, String filePath, String... otherValues) {
+    public static RequestBody fileForm(String fileFormKey,
+                                       String filePath,
+                                       final List<String> emptyKeyList, //可以为空的key
+                                       String... otherValues) {
         final MultipartBody.Builder builder = new MultipartBody.Builder();
 
         if (!TextUtils.isEmpty(filePath)) {
@@ -352,6 +385,19 @@ public class Http {
             @Override
             public void onRemove(String key) {
 
+            }
+
+            @Override
+            public boolean isKeyAllowEmpty(String key) {
+                if (emptyKeyList != null) {
+                    return emptyKeyList.contains(key);
+                }
+                return false;
+            }
+
+            @Override
+            public void onEmptyValue(String key, String value) {
+                builder.addFormDataPart(key, "");
             }
         }, otherValues);
 
@@ -396,5 +442,15 @@ public class Http {
         void onValue(String key, String value);
 
         void onRemove(String key);
+
+        /**
+         * key是否允许为空
+         */
+        boolean isKeyAllowEmpty(String key);
+
+        /**
+         * 当value为空时, 可以自行决定 设置 "" or null
+         */
+        void onEmptyValue(String key, String value);
     }
 }
